+++ src/core/fetch-css.ts
@@ -4,6 +4,7 @@ import { djinnjsOutDir } from './config';
 /**
  * Appends resources to the documents head if it hasn't already been loaded.
  * @param filenames - a filename `sting` or an array of `string` CSS filenames or a URL -- exclude the file path and extension if local
+ * @deprecated - slated for removal in v1.0.0
  */
 export function fetchCSS(filenames: string | Array<string>): Promise<{}> {
     return new Promise(resolve => {